refactor(telemetry): move hooks handler from otel_dashboard to telemetry router
