+++ observal-server/api/routes/telemetry.py
@@ -1,6 +1,8 @@
 import asyncio
+import json
 import logging
-import uuid
+import re
+import time as _time
 from datetime import UTC, datetime
 
 from fastapi import APIRouter, Depends, Header, Request
@@ -10,15 +12,13 @@ from models.user import User, UserRole
 from schemas.telemetry import (
     IngestBatch,
     IngestResponse,
-    TelemetryBatch,
     TelemetryStatusResponse,
 )
 from services.clickhouse import (
-    insert_agent_interaction,
+    _query,
     insert_otel_logs,
     insert_scores,
     insert_spans,
-    insert_tool_call,
     insert_traces,
     query_recent_events,
 )
@@ -30,6 +30,442 @@ router = APIRouter(prefix="/api/v1/telemetry", tags=["telemetry"])
 
 # Background tasks that must survive until completion (prevent GC)
 _background_tasks: set[asyncio.Task] = set()
+
+# ── Service name normalization ──
+# Normalize legacy ServiceName values to canonical IDE names.
+_SERVICE_NAME_MAP: dict[str, str] = {
+    "kiro-cli": "kiro",
+    "observal-hooks": "claude-code",
+    "observal-shim": "claude-code",
+    "copilot-cli": "copilot",
+    "github-copilot": "copilot",
+    "gemini-cli": "gemini",
+    "cursor-cli": "cursor",
+}
+
+# ── Kiro IDE session correlation ──
+_kiro_session_cache: dict[str, tuple[str, float]] = {}  # cwd -> (session_id, timestamp)
+_KIRO_SESSION_WINDOW = 1800  # 30 minutes
+
+# ── Per-IDE event name normalization maps ──
+
+_KIRO_TO_CC_EVENT = {
+    "agentSpawn": "SessionStart",
+    "userPromptSubmit": "UserPromptSubmit",
+    "preToolUse": "PreToolUse",
+    "promptSubmit": "UserPromptSubmit",  # Kiro IDE variant
+    "postToolUse": "PostToolUse",
+    "stop": "Stop",
+    "agentStop": "Stop",  # Kiro IDE variant
+}
+
+_GEMINI_TO_CC_EVENT = {
+    "SessionStart": "SessionStart",
+    "BeforeAgent": "SubagentStart",
+    "AfterAgent": "SubagentStop",
+    "BeforeTool": "PreToolUse",
+    "AfterTool": "PostToolUse",
+    "BeforeToolSelection": "PreToolUse",
+    "AfterModel": "Stop",
+    "SessionEnd": "Stop",
+}
+
+_CURSOR_TO_CC_EVENT = {
+    "preToolUse": "PreToolUse",
+    "postToolUse": "PostToolUse",
+    "sessionStart": "SessionStart",
+    "sessionEnd": "Stop",
+    "stop": "Stop",
+}
+
+_COPILOT_TO_CC_EVENT = {
+    "onPreToolUse": "PreToolUse",
+    "preToolUse": "PreToolUse",
+    "onPostToolUse": "PostToolUse",
+    "postToolUse": "PostToolUse",
+    "onUserPromptSubmitted": "UserPromptSubmit",
+    "onSessionStart": "SessionStart",
+    "sessionStart": "SessionStart",
+    "onSessionEnd": "Stop",
+    "sessionEnd": "Stop",
+    "onErrorOccurred": "StopFailure",
+    "subagent.started": "SubagentStart",
+    "subagent.completed": "SubagentStop",
+    "subagent.failed": "SubagentStop",
+    "subagent.selected": "SubagentStart",
+    "subagent.deselected": "SubagentStop",
+}
+
+# ── Per-IDE field normalization maps ──
+
+_KIRO_FIELD_MAP = {
+    "hookEventName": "hook_event_name",
+    "sessionId": "session_id",
+    "toolName": "tool_name",
+    "toolInput": "tool_input",
+    "toolResponse": "tool_response",
+    "toolUseId": "tool_use_id",
+    "agentId": "agent_id",
+    "agentType": "agent_type",
+    "stopReason": "stop_reason",
+    "permissionMode": "permission_mode",
+    "lastAssistantMessage": "last_assistant_message",
+    "userPrompt": "user_prompt",
+}
+
+_GEMINI_FIELD_MAP = {
+    "hookEventName": "hook_event_name",
+    "hook_event_name": "hook_event_name",
+    "sessionId": "session_id",
+    "session_id": "session_id",
+    "toolName": "tool_name",
+    "tool_name": "tool_name",
+    "toolInput": "tool_input",
+    "toolResponse": "tool_response",
+    "transcriptPath": "transcript_path",
+    "transcript_path": "transcript_path",
+}
+
+_CURSOR_FIELD_MAP = {
+    "hookEventName": "hook_event_name",
+    "hook_event_name": "hook_event_name",
+    "sessionId": "session_id",
+    "session_id": "session_id",
+    "toolName": "tool_name",
+    "tool_name": "tool_name",
+    "toolInput": "tool_input",
+    "tool_input": "tool_input",
+    "toolOutput": "tool_response",
+    "tool_output": "tool_response",
+    "toolUseId": "tool_use_id",
+    "tool_use_id": "tool_use_id",
+    "agentMessage": "agent_message",
+    "agent_message": "agent_message",
+}
+
+_COPILOT_FIELD_MAP = {
+    "hookEventName": "hook_event_name",
+    "hook_event_name": "hook_event_name",
+    "sessionId": "session_id",
+    "session_id": "session_id",
+    "toolName": "tool_name",
+    "tool_name": "tool_name",
+    "toolInput": "tool_input",
+    "tool_input": "tool_input",
+    "toolResponse": "tool_response",
+    "tool_response": "tool_response",
+    "toolCallId": "tool_use_id",
+    "toolUseId": "tool_use_id",
+    "tool_use_id": "tool_use_id",
+    "agentName": "agent_name",
+    "agentDisplayName": "agent_display_name",
+    "agentDescription": "agent_description",
+    "stopReason": "stop_reason",
+    "userPrompt": "user_prompt",
+}
+
+
+# ── Hook ingestion helpers ──
+
+
+def _truncate(s: str, max_len: int = 64000) -> str:
+    """Truncate a string to fit in ClickHouse without blowing up storage."""
+    if len(s) <= max_len:
+        return s
+    return s[:max_len] + f"\n... [truncated, {len(s)} total chars]"
+
+
+def _safe_json(obj: object) -> str:
+    """Serialize to JSON string, falling back to str()."""
+    if obj is None:
+        return ""
+    if isinstance(obj, str):
+        return obj
+    try:
+        return json.dumps(obj, default=str)
+    except Exception:
+        return str(obj)
+
+
+# ── IDE-specific extraction helpers ──
+
+
+def _extract_kiro(body: dict, hook_event: str, attrs: dict[str, str]) -> None:
+    """Extract Kiro-specific fields into *attrs*."""
+    tool_input_raw = body.get("tool_input")
+    tool_response_raw = body.get("tool_response")
+
+    if tool_input_raw is not None:
+        attrs["tool_input"] = _truncate(_safe_json(tool_input_raw))
+    if tool_response_raw is not None:
+        attrs["tool_response"] = _truncate(_safe_json(tool_response_raw))
+
+    if body.get("prompt") and not attrs.get("tool_input"):
+        attrs["tool_input"] = _truncate(str(body["prompt"]))
+    if body.get("assistant_response") and not attrs.get("tool_response"):
+        attrs["tool_response"] = _truncate(str(body["assistant_response"]))
+
+    if hook_event == "UserPromptSubmit":
+        prompt_text = body.get("prompt") or body.get("user_prompt") or ""
+        if prompt_text:
+            attrs["tool_input"] = _truncate(prompt_text)
+            attrs["prompt_length"] = str(len(prompt_text))
+        attrs["tool_name"] = "user_prompt"
+
+    if hook_event == "SessionStart":
+        prompt_text = body.get("prompt") or ""
+        if prompt_text:
+            attrs["tool_input"] = _truncate(prompt_text)
+        attrs["event.name"] = "hook_sessionstart"
+
+    if hook_event == "Stop":
+        if body.get("stop_reason"):
+            attrs["stop_reason"] = body["stop_reason"]
+        if body.get("assistant_response"):
+            attrs["tool_response"] = _truncate(str(body["assistant_response"]))
+
+    if hook_event == "PostToolUseFailure" and body.get("error"):
+        attrs["error"] = _truncate(str(body["error"]))
+
+    tool_name = attrs.get("tool_name", "")
+    if hook_event in ("PreToolUse", "PostToolUse") and tool_input_raw:
+        try:
+            ti = json.loads(tool_input_raw) if isinstance(tool_input_raw, str) else tool_input_raw
+        except (json.JSONDecodeError, TypeError):
+            ti = None
+        if isinstance(ti, dict):
+            if tool_name == "Agent":
+                if ti.get("subagent_type"):
+                    attrs["agent_type"] = ti["subagent_type"]
+                if ti.get("description"):
+                    attrs["agent_name"] = ti["description"]
+            elif tool_name == "Skill":
+                if ti.get("skill"):
+                    attrs["skill_name"] = ti["skill"]
+
+    for enriched_field in (
+        "input_tokens",
+        "output_tokens",
+        "turn_count",
+        "credits",
+        "tools_used",
+        "conversation_id",
+    ):
+        if body.get(enriched_field):
+            attrs[enriched_field] = str(body[enriched_field])
+
+
+def _extract_claude_code(body: dict, hook_event: str, attrs: dict[str, str]) -> None:
+    """Extract Claude Code-specific fields into *attrs*."""
+    tool_name = attrs.get("tool_name", "")
+    tool_input_raw = body.get("tool_input")
+    tool_response_raw = body.get("tool_response")
+
+    if tool_input_raw is not None:
+        attrs["tool_input"] = _truncate(_safe_json(tool_input_raw))
+    if tool_response_raw is not None:
+        attrs["tool_response"] = _truncate(_safe_json(tool_response_raw))
+
+    if hook_event == "UserPromptSubmit":
+        prompt_text = body.get("user_prompt") or body.get("prompt") or ""
+        if prompt_text:
+            attrs["tool_input"] = _truncate(prompt_text)
+            attrs["prompt_length"] = str(len(prompt_text))
+        attrs["tool_name"] = "user_prompt"
+
+    if hook_event == "SessionStart":
+        prompt_text = body.get("prompt") or ""
+        if prompt_text:
+            attrs["tool_input"] = _truncate(prompt_text)
+        attrs["event.name"] = "hook_sessionstart"
+        source = body.get("source", "")
+        if source:
+            attrs["session_source"] = source
+        if source in ("resume", "compact") or body.get("resume"):
+            attrs["session_resumed"] = "true"
+
+    if hook_event in ("PreToolUse", "PostToolUse") and tool_input_raw:
+        try:
+            ti = json.loads(tool_input_raw) if isinstance(tool_input_raw, str) else tool_input_raw
+        except (json.JSONDecodeError, TypeError):
+            ti = None
+        if isinstance(ti, dict):
+            if tool_name == "Agent":
+                if ti.get("subagent_type"):
+                    attrs["agent_type"] = ti["subagent_type"]
+                if ti.get("description"):
+                    attrs["agent_name"] = ti["description"]
+            elif tool_name == "Skill":
+                if ti.get("skill"):
+                    attrs["skill_name"] = ti["skill"]
+
+    if hook_event == "SubagentStart" and body.get("last_assistant_message"):
+        attrs["tool_input"] = _truncate(body["last_assistant_message"])
+    if hook_event == "SubagentStop" and body.get("last_assistant_message"):
+        attrs["tool_response"] = _truncate(body["last_assistant_message"])
+
+    if hook_event == "PostToolUseFailure" and body.get("error"):
+        attrs["error"] = _truncate(str(body["error"]))
+
+    if hook_event == "Stop":
+        if body.get("stop_reason"):
+            attrs["stop_reason"] = body["stop_reason"]
+        if tool_name == "assistant_response" and tool_response_raw:
+            attrs["event.name"] = "hook_assistant_response"
+        elif tool_name == "assistant_thinking" and tool_response_raw:
+            attrs["event.name"] = "hook_assistant_thinking"
+        if body.get("message_sequence") is not None:
+            attrs["message_sequence"] = str(body["message_sequence"])
+        if body.get("message_total") is not None:
+            attrs["message_total"] = str(body["message_total"])
+
+    if hook_event == "StopFailure":
+        if body.get("error"):
+            attrs["error"] = _truncate(str(body["error"]))
+        if body.get("stop_reason"):
+            attrs["stop_reason"] = body["stop_reason"]
+
+    if hook_event == "Notification":
+        if body.get("message"):
+            attrs["tool_response"] = _truncate(str(body["message"]))
+        if body.get("title"):
+            attrs["notification_title"] = str(body["title"])
+
+    if hook_event in ("TaskCreated", "TaskCompleted"):
+        for field in ("task_id", "task_subject", "task_status"):
+            if body.get(field):
+                attrs[field] = str(body[field])
+
+    if hook_event in ("PreCompact", "PostCompact") and body.get("summary"):
+        attrs["tool_response"] = _truncate(str(body["summary"]))
+
+    if hook_event in ("WorktreeCreate", "WorktreeRemove"):
+        if body.get("worktree_path"):
+            attrs["worktree_path"] = str(body["worktree_path"])
+        if body.get("branch"):
+            attrs["branch"] = str(body["branch"])
+
+    if hook_event in ("Elicitation", "ElicitationResult"):
+        for field in ("mcp_server_name", "message", "response", "elicitation_id"):
+            if body.get(field):
+                key = "tool_input" if field == "message" else ("tool_response" if field == "response" else field)
+                attrs[key] = _truncate(str(body[field]))
+
+
+def _extract_gemini(body: dict, hook_event: str, attrs: dict[str, str]) -> None:
+    """Extract Gemini CLI-specific fields into *attrs*."""
+    tool_input_raw = body.get("tool_input")
+    tool_response_raw = body.get("tool_response")
+
+    if tool_input_raw is not None:
+        attrs["tool_input"] = _truncate(_safe_json(tool_input_raw))
+    if tool_response_raw is not None:
+        attrs["tool_response"] = _truncate(_safe_json(tool_response_raw))
+
+    if hook_event == "SessionStart":
+        attrs["event.name"] = "hook_sessionstart"
+
+    if hook_event in ("SubagentStart", "SubagentStop") and body.get("additional_context"):
+        field = "tool_input" if hook_event == "SubagentStart" else "tool_response"
+        attrs[field] = _truncate(str(body["additional_context"]))
+
+    if hook_event == "PreToolUse":
+        tool_name = body.get("tool_name") or ""
+        if tool_name:
+            attrs["tool_name"] = tool_name
+        llm_req = body.get("llm_request")
+        if isinstance(llm_req, dict):
+            messages = llm_req.get("messages") or []
+            if messages:
+                last_user = None
+                for m in reversed(messages):
+                    if m.get("role") == "user":
+                        last_user = m
+                        break
+                if last_user:
+                    content = last_user.get("content", "")
+                    if isinstance(content, str) and not attrs.get("tool_input"):
+                        attrs["tool_input"] = _truncate(content)
+
+    if hook_event == "Stop" and body.get("stop_reason"):
+        attrs["stop_reason"] = body["stop_reason"]
+
+    if body.get("model"):
+        attrs["model"] = str(body["model"])
+
+    for enriched_field in ("input_tokens", "output_tokens", "total_tokens"):
+        if body.get(enriched_field):
+            attrs[enriched_field] = str(body[enriched_field])
+
+
+def _extract_cursor(body: dict, hook_event: str, attrs: dict[str, str]) -> None:
+    """Extract Cursor-specific fields into *attrs*."""
+    tool_input_raw = body.get("tool_input")
+    tool_response_raw = body.get("tool_response") or body.get("tool_output")
+
+    if tool_input_raw is not None:
+        attrs["tool_input"] = _truncate(_safe_json(tool_input_raw))
+    if tool_response_raw is not None:
+        attrs["tool_response"] = _truncate(_safe_json(tool_response_raw))
+
+    if hook_event == "SessionStart":
+        attrs["event.name"] = "hook_sessionstart"
+
+    if hook_event == "PreToolUse" and body.get("agent_message"):
+        attrs["agent_message"] = _truncate(str(body["agent_message"]))
+
+    if hook_event in ("PreToolUse", "PostToolUse") and body.get("duration"):
+        attrs["duration_ms"] = str(body["duration"])
+
+    if hook_event == "Stop" and body.get("stop_reason"):
+        attrs["stop_reason"] = body["stop_reason"]
+
+    if body.get("model"):
+        attrs["model"] = str(body["model"])
+
+
+def _extract_copilot(body: dict, hook_event: str, attrs: dict[str, str]) -> None:
+    """Extract GitHub Copilot-specific fields into *attrs*."""
+    tool_input_raw = body.get("tool_input")
+    tool_response_raw = body.get("tool_response")
+
+    if tool_input_raw is not None:
+        attrs["tool_input"] = _truncate(_safe_json(tool_input_raw))
+    if tool_response_raw is not None:
+        attrs["tool_response"] = _truncate(_safe_json(tool_response_raw))
+
+    if hook_event == "SessionStart":
+        attrs["event.name"] = "hook_sessionstart"
+        if body.get("source"):
+            attrs["session_source"] = str(body["source"])
+
+    if hook_event == "UserPromptSubmit":
+        prompt_text = body.get("user_prompt") or body.get("prompt") or ""
+        if prompt_text:
+            attrs["tool_input"] = _truncate(prompt_text)
+            attrs["prompt_length"] = str(len(prompt_text))
+        attrs["tool_name"] = "user_prompt"
+
+    if hook_event in ("SubagentStart", "SubagentStop"):
+        if body.get("agent_name"):
+            attrs["agent_name"] = body["agent_name"]
+        if body.get("agent_display_name"):
+            attrs["agent_name"] = body["agent_display_name"]
+        if body.get("agent_description"):
+            attrs["agent_type"] = body["agent_description"][:100]
+        if body.get("tool_use_id"):
+            attrs["tool_use_id"] = body["tool_use_id"]
+        if hook_event == "SubagentStop" and body.get("error"):
+            attrs["error"] = _truncate(str(body["error"]))
+
+    if hook_event == "Stop" and (body.get("stop_reason") or body.get("reason")):
+        attrs["stop_reason"] = body.get("stop_reason") or body.get("reason", "")
+
+    if hook_event == "StopFailure" and body.get("error"):
+        attrs["error"] = _truncate(str(body["error"]))
+
+    if body.get("model"):
+        attrs["model"] = str(body["model"])
 
 
 # Shim span type → otel_logs event.name mapping
@@ -286,59 +722,6 @@ async def ingest(
     return IngestResponse(ingested=ingested, errors=errors)
 
 
-@router.post("/events")
-async def ingest_events(
-    batch: TelemetryBatch,
-    current_user: User = Depends(require_role(UserRole.user)),
-):
-    now = datetime.now(UTC).strftime("%Y-%m-%d %H:%M:%S.%f")[:-3]
-    ingested = 0
-    errors = 0
-
-    for tc in batch.tool_calls:
-        try:
-            await insert_tool_call(
-                {
-                    "event_id": str(uuid.uuid4()),
-                    "timestamp": now,
-                    "mcp_server_id": tc.mcp_server_id,
-                    "tool_name": tc.tool_name,
-                    "input_params": redact_secrets(tc.input_params) if tc.input_params else tc.input_params,
-                    "response": redact_secrets(tc.response) if tc.response else tc.response,
-                    "latency_ms": tc.latency_ms,
-                    "status": tc.status,
-                    "user_action": tc.user_action,
-                    "session_id": tc.session_id,
-                    "user_id": str(current_user.id),
-                    "ide": tc.ide,
-                }
-            )
-            ingested += 1
-        except Exception:
-            errors += 1
-
-    for ai in batch.agent_interactions:
-        try:
-            await insert_agent_interaction(
-                {
-                    "event_id": str(uuid.uuid4()),
-                    "timestamp": now,
-                    "agent_id": ai.agent_id,
-                    "session_id": ai.session_id,
-                    "tool_calls": ai.tool_calls,
-                    "user_action": ai.user_action,
-                    "latency_ms": ai.latency_ms,
-                    "user_id": str(current_user.id),
-                    "ide": ai.ide,
-                }
-            )
-            ingested += 1
-        except Exception:
-            errors += 1
-
-    return {"ingested": ingested, "errors": errors}
-
-
 @router.get("/status", response_model=TelemetryStatusResponse)
 async def telemetry_status(current_user: User = Depends(require_role(UserRole.admin))):
     counts = await query_recent_events(60)
@@ -349,103 +732,215 @@ async def telemetry_status(current_user: User = Depends(require_role(UserRole.ad
     )
 
 
-# Kiro CLI and Copilot CLI use camelCase event names; normalize to PascalCase
-_KIRO_EVENT_MAP = {
-    # Kiro events
-    "agentSpawn": "SessionStart",
-    "userPromptSubmit": "UserPromptSubmit",
-    "promptSubmit": "UserPromptSubmit",
-    "preToolUse": "PreToolUse",
-    "postToolUse": "PostToolUse",
-    "stop": "Stop",
-    "agentStop": "Stop",
-    # Copilot CLI events
-    "sessionStart": "SessionStart",
-    "sessionEnd": "Stop",
-    "userPromptSubmitted": "UserPromptSubmit",
-    "errorOccurred": "StopFailure",
-}
-
-
 @router.post("/hooks")
-async def ingest_hook(request: Request, current_user: User = Depends(require_role(UserRole.user))):
-    """Ingest raw hook JSON from Claude Code/Kiro/Copilot CLI."""
-    project_id = get_project_id(current_user)
-    body = await request.json()
+async def ingest_hook(request: Request):
+    """Ingest hook events from IDE agents and store in otel_logs.
 
-    # Normalize Kiro camelCase field names to snake_case
-    if "hookEventName" in body and "hook_event_name" not in body:
-        body["hook_event_name"] = body["hookEventName"]
-    if "toolName" in body and "tool_name" not in body:
-        body["tool_name"] = body["toolName"]
-    if "toolInput" in body and "tool_input" not in body:
-        body["tool_input"] = body["toolInput"]
-    if "toolResponse" in body and "tool_response" not in body:
-        body["tool_response"] = body["toolResponse"]
-    if "sessionId" in body and "session_id" not in body:
-        body["session_id"] = body["sessionId"]
+    Supports Claude Code, Kiro, Gemini CLI, Cursor, and GitHub Copilot.
+    Each IDE's camelCase fields and event names are normalized to a
+    canonical snake_case/PascalCase schema before storage.
 
-    now = datetime.now(UTC).strftime("%Y-%m-%d %H:%M:%S.%f")[:-3]
-    raw_event = body.get("hook_event_name", "unknown")
-    hook_event_name = _KIRO_EVENT_MAP.get(raw_event, raw_event)
-    tool_name = body.get("tool_name", "")
-    if tool_name in ("assistant_response", "assistant_thinking"):
-        span_type = f"hook_{tool_name}"
-    elif hook_event_name == "PostToolUse":
-        span_type = "hook_exec"
+    Intentionally unauthenticated — CLI hooks can't easily carry auth tokens.
+    Supports ECIES-encrypted payloads via the ``X-Observal-Encrypted`` header.
+    """
+    encrypted_header = request.headers.get("X-Observal-Encrypted")
+    if encrypted_header == "ecies-p256":
+        raw_body = await request.body()
+        from services.crypto import get_key_manager
+
+        km = get_key_manager()
+        decrypted_json = km.decrypt_payload(raw_body)
+        body = json.loads(decrypted_json)
     else:
-        span_type = f"hook_{hook_event_name.lower()}"
-    row = {
-        "span_id": str(uuid.uuid4()),
-        "trace_id": body.get("session_id", str(uuid.uuid4())),
-        "parent_span_id": None,
-        "project_id": project_id,
-        "mcp_id": None,
-        "agent_id": None,
-        "user_id": str(current_user.id),
-        "type": span_type,
-        "name": body.get("tool_name", "hook"),
-        "method": "",
-        "input": body.get("tool_input"),
-        "output": body.get("tool_response"),
-        "error": None,
-        "start_time": now,
-        "end_time": now,
-        "latency_ms": None,
-        "status": "success",
-        "ide": body.get("service_name", ""),
-        "environment": "default",
-        "metadata": {},
-        "token_count_input": None,
-        "token_count_output": None,
-        "token_count_total": None,
-        "cost": None,
-        "cpu_ms": None,
-        "memory_mb": None,
-        "hop_count": None,
-        "entities_retrieved": None,
-        "relationships_used": None,
-        "retry_count": None,
-        "tools_available": None,
-        "tool_schema_valid": None,
-        "container_id": None,
-        "exit_code": None,
-        "network_bytes_in": None,
-        "network_bytes_out": None,
-        "disk_read_bytes": None,
-        "disk_write_bytes": None,
-        "oom_killed": None,
-        "query_interface": None,
-        "relevance_score": None,
-        "chunks_returned": None,
-        "embedding_latency_ms": None,
-        "hook_event": hook_event_name,
-        "hook_scope": None,
-        "hook_action": None,
-        "hook_blocked": None,
-        "variables_provided": None,
-        "template_tokens": None,
-        "rendered_tokens": None,
+        body = await request.json()
+    now = datetime.now(UTC).strftime("%Y-%m-%d %H:%M:%S.%f")[:-3]
+
+    # ── Detect IDE early so we pick the right normalization maps ──
+    raw_service = body.get("service_name") or body.get("serviceName") or "claude-code"
+    service_name = _SERVICE_NAME_MAP.get(raw_service, raw_service)
+
+    # ── Normalize camelCase fields to snake_case (per-IDE map) ──
+    field_map = {
+        "kiro": _KIRO_FIELD_MAP,
+        "gemini": _GEMINI_FIELD_MAP,
+        "cursor": _CURSOR_FIELD_MAP,
+        "copilot": _COPILOT_FIELD_MAP,
+    }.get(service_name, _KIRO_FIELD_MAP)
+    normalized: dict = {}
+    for key, value in body.items():
+        normalized[field_map.get(key, key)] = value
+    body = normalized
+
+    # ── Normalize event names to canonical PascalCase (per-IDE map) ──
+    raw_event = body.get("hook_event_name", "unknown")
+    event_map = {
+        "kiro": _KIRO_TO_CC_EVENT,
+        "gemini": _GEMINI_TO_CC_EVENT,
+        "cursor": _CURSOR_TO_CC_EVENT,
+        "copilot": _COPILOT_TO_CC_EVENT,
+    }.get(service_name, {})
+    hook_event = event_map.get(raw_event, raw_event)
+
+    # Kiro/Cursor postToolUse with tool_response.success=false → PostToolUseFailure
+    if hook_event == "PostToolUse":
+        tool_resp = body.get("tool_response")
+        if isinstance(tool_resp, dict) and tool_resp.get("success") is False:
+            hook_event = "PostToolUseFailure"
+            result = tool_resp.get("result", "")
+            if result and not body.get("error"):
+                body["error"] = _truncate(str(result))
+
+    body["hook_event_name"] = hook_event
+
+    session_id = body.get("session_id", "")
+    tool_name = body.get("tool_name", "")
+
+    # ── Kiro IDE session correlation ──
+    # $PPID differs per hook invocation in IDE context. Correlate by cwd.
+    cwd = body.get("cwd", "")
+    is_kiro = service_name == "kiro"
+    is_kiro_ppid = bool(re.match(r"^kiro-\d+$", session_id))
+
+    # Fallback: if a Kiro event arrives with no session_id, synthesize one from cwd
+    if is_kiro and not session_id and cwd:
+        import hashlib
+
+        session_id = f"kiro-{hashlib.sha256(cwd.encode()).hexdigest()[:12]}"
+        is_kiro_ppid = True  # treat it like a PPID session for caching below
+
+    if is_kiro_ppid and cwd:
+        now_ts = _time.monotonic()
+        cached = _kiro_session_cache.get(cwd)
+        if cached and (now_ts - cached[1]) < _KIRO_SESSION_WINDOW:
+            session_id = cached[0]
+        else:
+            _kiro_session_cache[cwd] = (session_id, now_ts)
+        body["session_id"] = session_id
+
+    # Build the attributes map that the frontend already reads
+    attrs: dict[str, str] = {
+        "session.id": session_id,
+        "event.name": f"hook_{hook_event.lower()}",
+        "hook_event": hook_event,
+        "tool_name": tool_name,
+        "source": "hook",
     }
-    await insert_spans([row])
-    return {"ingested": 1}
+
+    # ── Agent attribution ──
+    if body.get("agent_id"):
+        attrs["agent_id"] = body["agent_id"]
+    if body.get("agent_type"):
+        attrs["agent_type"] = body["agent_type"]
+    if body.get("agent_name"):
+        attrs["agent_name"] = body["agent_name"]
+    if body.get("model"):
+        attrs["model"] = body["model"]
+
+    # ── User identity (from Observal login, injected by CLI) ──
+    user_id = body.get("user_id") or request.headers.get("x-observal-user-id") or ""
+    if user_id:
+        attrs["user.id"] = user_id
+
+    user_name = body.get("user_name") or request.headers.get("x-observal-username") or ""
+    if user_name:
+        attrs["user.name"] = user_name
+
+    # ── IDE-specific extraction ──
+    _ide_extractors = {
+        "kiro": _extract_kiro,
+        "gemini": _extract_gemini,
+        "cursor": _extract_cursor,
+        "copilot": _extract_copilot,
+    }
+    extractor = _ide_extractors.get(service_name, _extract_claude_code)
+    extractor(body, hook_event, attrs)
+
+    # Extra context fields (present on most events, all IDEs)
+    if body.get("tool_use_id"):
+        attrs["tool_use_id"] = body["tool_use_id"]
+    if body.get("cwd"):
+        attrs["cwd"] = body["cwd"]
+    if body.get("permission_mode"):
+        attrs["permission_mode"] = body["permission_mode"]
+
+    # Build the Body as a readable summary
+    agent_prefix = f"[{attrs.get('agent_type', '')}] " if attrs.get("agent_id") else ""
+    if hook_event in ("PostToolUse", "PreToolUse"):
+        body_text = f"{agent_prefix}{hook_event}: {tool_name}"
+    elif hook_event == "PostToolUseFailure":
+        body_text = f"{agent_prefix}ToolFailure: {tool_name}"
+    elif hook_event == "UserPromptSubmit":
+        prompt_preview = (attrs.get("tool_input") or "")[:100]
+        body_text = f"Prompt: {prompt_preview}"
+    elif hook_event in ("SubagentStop", "SubagentStart"):
+        body_text = f"{hook_event}: {attrs.get('agent_type', 'unknown')}"
+    elif hook_event in ("Elicitation", "ElicitationResult"):
+        body_text = f"{hook_event}: {body.get('mcp_server_name', 'unknown')}"
+    elif hook_event == "Stop" and tool_name == "assistant_thinking":
+        seq = attrs.get("message_sequence", "")
+        total = attrs.get("message_total", "")
+        seq_label = f" [{seq}/{total}]" if seq and total else ""
+        preview = (attrs.get("tool_response") or "")[:100]
+        body_text = f"Thinking{seq_label}: {preview}"
+    elif hook_event == "Stop" and (tool_name == "assistant_response" or body.get("assistant_response")):
+        seq = attrs.get("message_sequence", "")
+        total = attrs.get("message_total", "")
+        seq_label = f" [{seq}/{total}]" if seq and total else ""
+        preview = (attrs.get("tool_response") or "")[:100]
+        body_text = f"Response{seq_label}: {preview}"
+    elif hook_event == "Stop":
+        body_text = f"Stop: {body.get('stop_reason', 'end_turn')}"
+    elif hook_event == "StopFailure":
+        body_text = f"StopFailure: {attrs.get('error', 'unknown')[:80]}"
+    elif hook_event == "SessionStart":
+        source_label = ""
+        if attrs.get("session_source") == "compact":
+            source_label = " (continued)"
+        elif attrs.get("session_resumed") == "true":
+            source_label = " (resumed)"
+        prompt_preview = (attrs.get("tool_input") or "")[:100]
+        body_text = f"SessionStart{source_label}: {prompt_preview}" if prompt_preview else f"SessionStart{source_label}"
+    elif hook_event == "Notification":
+        body_text = f"Notification: {attrs.get('notification_title', '')}"
+    elif hook_event in ("TaskCreated", "TaskCompleted"):
+        body_text = f"{hook_event}: {attrs.get('task_subject', '')[:60]}"
+    elif hook_event in ("PreCompact", "PostCompact"):
+        body_text = f"{hook_event}"
+    elif hook_event in ("WorktreeCreate", "WorktreeRemove"):
+        body_text = f"{hook_event}: {attrs.get('branch', '')}"
+    else:
+        body_text = f"{agent_prefix}hook: {hook_event}"
+
+    # ── Redact secrets from user-content fields before storage ──
+    for _redact_field in ("tool_input", "tool_response", "error"):
+        if _redact_field in attrs:
+            attrs[_redact_field] = redact_secrets(attrs[_redact_field])
+    body_text = redact_secrets(body_text)
+
+    # INSERT into otel_logs using JSONEachRow (safe against injection)
+    row = {
+        "Timestamp": now,
+        "Body": body_text,
+        "LogAttributes": attrs,
+        "ServiceName": service_name,
+        "SeverityText": "INFO",
+        "SeverityNumber": 9,
+    }
+    sql = "INSERT INTO otel_logs (Timestamp, Body, LogAttributes, ServiceName, SeverityText, SeverityNumber) FORMAT JSONEachRow"
+
+    try:
+        r = await _query(sql, data=json.dumps(row, default=str))
+        if r.status_code != 200:
+            logger.warning("hook_insert_failed", extra={"status_code": r.status_code, "response": r.text[:200]})
+            return {"ingested": 0, "error": "insert failed"}
+    except Exception as e:
+        logger.warning("hook_insert_failed", extra={"error": str(e)})
+        return {"ingested": 0, "error": str(e)}
+
+    # Notify subscribers (fire-and-forget — don't block the response)
+    if session_id:
+        task = asyncio.create_task(publish("sessions:updated", {"session_id": session_id, "event_name": hook_event}))
+        _background_tasks.add(task)
+        task.add_done_callback(_background_tasks.discard)
+
+    return {"ingested": 1, "session_id": session_id, "event": hook_event}
